--- src/app/faqs/page.tsx
@@ -1,12 +1,0 @@
-import Faq from '@/components/faq/Faq';
-import React from 'react';
-
-const page = () => {
-  return (
-    <div>
-      <Faq />
-    </div>
-  );
-};
-
-export default page;
