+++ src/app/schedule/styles/ScheduleItem.module.scss
@@ -5,7 +5,7 @@
   align-items: stretch;
   gap: 2rem;
   margin-bottom: 1rem;
-  width: 60vw;
+  width: 50vw;
   height: auto;
 
   @include media('<=phone') {
