--- src/app/speakers/mockData/SpeakerData.ts
@@ -1,197 +1,0 @@
-export interface Speaker {
-  career: string;
-  name: string;
-  image: string;
-  topic: string;
-  date: string;
-  bio: string;
-  socials: {
-    twitter: string;
-    linkdin: string;
-    instagram: string;
-    facebook: string;
-  };
-}
-
-export const fullSpeakerData: Speaker[] = [
-  {
-    career: 'senior software engineer',
-    name: 'Adora Hack',
-    image: '/images/speakers/image-13.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    career: 'senior software engineer',
-    name: 'Babatunde A.K.A Lala',
-    image: '/images/speakers/image-14.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Cassandra Casanova',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-15.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Dayo The Striker',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-17.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Eve Encryption',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-19.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Frankie Function',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-20.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Grace Grid',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-23.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Hank Hash',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-14.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-  {
-    name: 'Ian interface',
-    career: 'senior software engineer',
-    image: '/images/speakers/image-15.png',
-    topic: `Building a Secure Interchain Messa
-ging dApp: A Technical Deep Building a Secure Interchain`,
-    date: 'Saturday, 1:50 PM - 2:30 PM',
-    bio: `Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    
-    Building a Secure Interchain Messaging dApp: A Technical Deep Building a Secure Interchain 
-
-    Lorem ipsum dolor sit amet consectetur. Elit porta nunc morbi risus egestas sit massa lacinia facilisis. Molestie tristique iaculis tellus sem. Mi risus mauris aenean nibh turpis interdum ultrices sem quisque.
-    `,
-    socials: {
-      twitter: 'twitter',
-      linkdin: 'linkdin',
-      instagram: 'instagram',
-      facebook: 'facebook',
-    },
-  },
-];
