--- src/app/speakers/mockData/mock.ts
@@ -1,42 +1,0 @@
-const speakersSlideData = [
-  {
-    image: '/images/speakers/image-13.png',
-    name: 'Adewale Abati',
-  },
-  {
-    image: '/images/speakers/image-14.png',
-    name: 'Adewale Abati',
-  },
-  {
-    image: '/images/speakers/image-15.png',
-    name: 'Adewale Abati',
-  },
-  {
-    image: '/images/speakers/image-17.png',
-    name: 'Adewale Abati',
-  },
-  {
-    image: '/images/speakers/image-20.png',
-    name: 'Adewale Abati',
-  },
-  {
-    image: '/images/speakers/image-23.png',
-    name: 'Adewale Abati',
-  },
-];
-
-const categories = [
-  'All',
-  'Product design',
-  'Front end',
-  'AI',
-  'Back end',
-  'Technical writing',
-  'Data analysis',
-  'Blockchain',
-  'Dev ops',
-  'Cloud computing',
-  'Brand design',
-];
-
-export { speakersSlideData, categories };
