--- src/app/speakers/styles/DaysAndFilter.module.scss
@@ -1,112 +1,0 @@
-@import '/src/styles/core/variables';
-@import '/src/styles/utils/breakpoints';
-@import '/src/styles/core/typography';
-
-.DayFilterWrapper {
-  background: $yellow-primary-100;
-  padding: 1.5rem 19rem;
-  padding-bottom: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  button {
-    background: inherit;
-    cursor: pointer;
-    transition: background-color ease 0.2s;
-    border-radius: 2em;
-    color: black;
-  }
-
-  .DayBtnsContainer {
-    display: flex;
-    gap: 1.5rem;
-    align-items: center;
-    justify-content: center;
-    padding: 1.5rem;
-
-    .dayBtn {
-      @include body2-medium();
-      padding: 1rem 7.2rem;
-      border: 1px solid $grey-10;
-      text-wrap: nowrap;
-
-      @include media('<=tablet') {
-        @include body3-medium();
-        padding: 1rem;
-      }
-
-      @include media('<=400px') {
-        @include body4-medium();
-        padding: 0.8rem;
-      }
-    }
-    .activeDayBtn {
-      background-color: $core-yellow;
-    }
-
-    @include media('<=desktop') {
-      padding: 0 1.6rem;
-      gap: 0.8rem;
-    }
-
-    @include media('<=tablet') {
-      gap: 0.5rem;
-    }
-
-    @include media('<=phone') {
-      padding: 0 1rem;
-    }
-  }
-
-  .FilterBtnsContainer {
-    padding: 3rem 0;
-
-    .FilterBtnsFlex {
-      display: flex;
-      row-gap: 1rem;
-      column-gap: 1.5rem;
-      flex-wrap: wrap;
-      width: fit-content;
-      justify-content: center;
-
-      .filterBtn {
-        @include body2-medium();
-        padding: 0.5rem 1.5rem;
-        border: 1px solid $grey-60;
-        color: $grey-40;
-        text-wrap: nowrap;
-        width: fit-content;
-      }
-      .activeFilterBtn {
-        border: 1px solid $blue-primary-40;
-        background-color: $core-green;
-        color: black;
-      }
-
-      @include media('<=desktop') {
-        display: grid;
-        grid-auto-flow: column;
-        gap: 0.5rem;
-        width: max-content;
-        .filterBtn {
-          display: block;
-        }
-      }
-    }
-
-    @include media('<=desktop') {
-      width: 100vw;
-      overflow-x: scroll;
-      overscroll-behavior-inline: contain;
-      -webkit-overflow-scrolling: touch;
-      padding: 1.5rem 2rem;
-    }
-  }
-
-  @include media('<=tablet') {
-    padding: 1rem 0;
-    padding-top: 4rem;
-    gap: 0rem;
-  }
-}
