+++ src/styles/index.scss
@@ -16,6 +16,7 @@
 @import 'components/faq';
 @import 'components/faq-accordion';
 @import 'components/community';
+@import 'pages/schedule';
 
 @import 'components/subscribe';
